ResponseEntity

+++ src/main/java/pl/zmt/reader/entities/CompressorController.java
@@ -2,8 +2,10 @@ package pl.zmt.reader.entities;
 
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
 
 import java.util.Optional;
 
@@ -16,25 +18,34 @@ public class CompressorController {
     CompressorRepository compressorRepository;
 
     @GetMapping
-    public @ResponseBody
-    Iterable<Compressor> getAllCompressors() {
-        return compressorRepository.findAll();
+    public ResponseEntity getAllCompressors() {
+        Iterable<Compressor> compressors = compressorRepository.findAll();
+        return ResponseEntity.ok(compressors);
     }
 
     @GetMapping("/{id}")
-    public @ResponseBody
-    Optional<Compressor> getCompressorById(@PathVariable Long id) {
-        return compressorRepository.findById(id);
+    public ResponseEntity compressorById (@PathVariable Long id) {
+        Optional<Compressor> compressorById = compressorRepository.findById(id);
+        return ResponseEntity.ok(compressorById);
     }
 
     @PostMapping("/add")
-    public Compressor addCompressor(@RequestBody Compressor compressor) {
-        return compressorRepository.save(compressor);
+    public ResponseEntity addCompressor(@RequestBody Compressor compressor) {
+        Compressor savedCompressor = compressorRepository.save(compressor);
+        return ResponseEntity.ok(savedCompressor);
     }
 
     @DeleteMapping("/{id}")
-    public void delCompressor(@PathVariable Long id) {
-        compressorRepository.deleteById(id);
+    public ResponseEntity delCompressor(@PathVariable Long id) {
+
+        boolean isPresent = compressorRepository.existsById(id);
+
+        if(isPresent) {
+            compressorRepository.deleteById(id);
+            return new ResponseEntity(id, HttpStatus.OK);
+        }
+
+        return new ResponseEntity("not found", HttpStatus.NOT_FOUND);
     }
 
 }
